code cleanup

+++ core/src/net/pzdcrp/Aselia/SuperPizdatiyShader.java
@@ -1,37 +1,19 @@
 package net.pzdcrp.Aselia;
 
-import java.nio.ByteBuffer;
-import java.nio.FloatBuffer;
-import java.nio.ShortBuffer;
 import java.util.List;
 import java.util.concurrent.CopyOnWriteArrayList;
 
 import com.badlogic.gdx.Gdx;
-import com.badlogic.gdx.graphics.Camera;
-import com.badlogic.gdx.graphics.Color;
-import com.badlogic.gdx.graphics.GL20;
-import com.badlogic.gdx.graphics.GLTexture;
-import com.badlogic.gdx.graphics.Pixmap;
-import com.badlogic.gdx.graphics.Pixmap.Format;
-import com.badlogic.gdx.graphics.Texture;
-import com.badlogic.gdx.graphics.g3d.Attribute;
 import com.badlogic.gdx.graphics.g3d.Attributes;
 import com.badlogic.gdx.graphics.g3d.Renderable;
 import com.badlogic.gdx.graphics.g3d.Shader;
 import com.badlogic.gdx.graphics.g3d.shaders.DefaultShader;
-import com.badlogic.gdx.graphics.g3d.shaders.BaseShader.Uniform;
 import com.badlogic.gdx.graphics.g3d.utils.BaseShaderProvider;
-import com.badlogic.gdx.graphics.g3d.utils.DefaultShaderProvider;
-import com.badlogic.gdx.graphics.g3d.utils.RenderContext;
-import com.badlogic.gdx.graphics.glutils.ShaderProgram;
-import com.badlogic.gdx.math.MathUtils;
-import com.badlogic.gdx.math.Matrix4;
 import com.badlogic.gdx.math.Vector2;
 import com.badlogic.gdx.math.Vector3;
 import com.badlogic.gdx.utils.GdxRuntimeException;
 
 import net.pzdcrp.Aselia.utils.GameU;
-import net.pzdcrp.Aselia.utils.MathU;
 
 public class SuperPizdatiyShader extends BaseShaderProvider {
     //public int test;
@@ -39,15 +21,15 @@ public class SuperPizdatiyShader extends BaseShaderProvider {
     public int stage = 0;
     public float skylightlevel = 1;
     public Vector3 pos = Vector3.Zero;
-    
+
     public void newstage() {
     	stage++;
     }
-    
+
     public void end() {
     	stage = 0;
     }
-    
+
     @Override
 	public Shader getShader (Renderable renderable) {
     	Shader suggestedShader = renderable.shader;
@@ -61,7 +43,7 @@ public class SuperPizdatiyShader extends BaseShaderProvider {
 		shaders.add(shader);
 		return shader;
 	}
- 
+
     @Override
     protected Shader createShader(Renderable renderable) {
     	if (renderable.userData != null) {
@@ -97,7 +79,7 @@ public class SuperPizdatiyShader extends BaseShaderProvider {
 
 class ChunkModelShader extends DefaultShader {
 	private SuperPizdatiyShader s;
-	private final int 
+	private final int
 	screensize = register(new Uniform("screensize")),
 	campos = register(new Uniform("campos"));
 
@@ -105,10 +87,10 @@ class ChunkModelShader extends DefaultShader {
 		super(renderable, config);
 		this.s = s;
 	}
-	
+
 	@Override
 	public void render (Renderable renderable, Attributes combinedAttributes) {
-		set(screensize, new Vector2((float) Gdx.graphics.getWidth(), (float) Gdx.graphics.getHeight()));
+		set(screensize, new Vector2(Gdx.graphics.getWidth(), Gdx.graphics.getHeight()));
 		set(campos, s.pos);
 		super.render(renderable, combinedAttributes);
 	}
@@ -122,7 +104,7 @@ class SkyModelShader extends DefaultShader {
 		super(renderable, config);
 		this.s = s;
 	}
-	
+
 	@Override
 	public void render (Renderable renderable, Attributes combinedAttributes) {
 		set(lightlevel, s.skylightlevel);
@@ -138,7 +120,7 @@ class ItemEntityShader extends DefaultShader {
 		super(renderable, config);
 		this.s = s;
 	}
-	
+
 	@Override
 	public void render (Renderable renderable, Attributes combinedAttributes) {
 		float f = (float) ((Object[])renderable.userData)[1];
